whirlwind_caravan/zaqar_based/requirements.txt
python-zaqarclient
